+++ SilKit/source/config/ParticipantConfiguration.hpp
@@ -231,6 +231,8 @@ struct Middleware
     bool registryAsFallbackProxy{ true };
     //! By default, requesting connection of other participants, and honoring these requests by other participants is enabled.
     bool experimentalRemoteParticipantConnection{ true };
+    //! Timeout for individual connection attempts (TCP, Local-Domain) and handshakes.
+    double connectTimeoutSeconds{5.0};
 };
 
 // ================================================================================